Added a bootstrap button w/history state change
Info about the site (filler text right now) is hidden. Clicking the
'about' button on the home page will show that text and add that action
to the history stack, such that the url will change to a #about page and
that clicking back will instead return to the main index page and hide
the about text again.